Address PR comments
* Refactor to standalone module
* Don't depend on evaluation package
* Remove everything but the Binary-Artifact
