+++ attestor/testdata/policy-binauthz-missingparam.yaml
@@ -14,20 +14,3 @@
 # PreventBinaryArtifacts : set to true to require that this project's SCM repo is
 # free of binary artifacts
 preventBinaryArtifacts: true
-
-# AllowedBinaryArtifacts : List of binary artifact paths to ignore
-# when checking for binary artifacts in a repo
-allowedBinaryArtifacts: []
-# List of allowed binary artifact paths as strings
-
-# EnsureNoVulnerabilities : set to true to require that this project is free
-# of vulnerabilities, as discovered from the OSV service
-ensureNoVulnerabilities: true
-
-# EnsurePinnedDependencies : set to true to require that this project pin dependencies
-# by hash/commit SHA
-ensurePinnedDependencies: true
-
-# EnsureCodeReviewed : set to true to require that the most recent commits in
-# this project have gone through a code review process
-ensureCodeReviewed: true
